Constrain WMIC query for local users
Without this commit, the WMIC command that pulls SIDs for mapping will
pull every SID from a Domain if the node is joined. For large domains,
this can be a very lengthy and expensive operation. According to the
comments and method name, this WMIC query is meant to obtain only the
mapping used for local users.

This commit constrains the WMIC query to the domain of `%computername%`
and thus only the local users. This should significantly speed up the
runtime of this module.