Use ENV variable for machine name in WMIC
Without this commit, the WMIC execution that needs to be constrained to the
local node is broken. It is using a string that would have expanded
correctly in a command prompt, but when called from the WIN32/Process
library in Ruby is taken literally. This causes the WMIC call to be
broken.

This commit corrects the issue by accessing the local environment
variable for the "COMPUTERNAME". This ensures that the correct value can
actually be used and interpolated for the constraining of the WMIC
query. On large domains a significant performance increase was
eperienced.